+++ lib/src/exceptions/blueprint_exception.dart
@@ -1,0 +1,8 @@
+class BlueprintException implements Exception {
+  final String message;
+
+  const BlueprintException(this.message);
+
+  @override
+  String toString() => 'BlueprintException: $message';
+}