+++ lib/src/exceptions/dependency_exception.dart
@@ -1,8 +1,0 @@
-class DependencyException implements Exception {
-  final String message;
-
-  const DependencyException(this.message);
-
-  @override
-  String toString() => 'DependencyException: $message';
-}