+++ lib/src/interfaces/i_blueprint.dart
@@ -1,7 +1,7 @@
-import 'i_base_dependency.dart';
+import 'i_base_blueprint.dart';
 import 'i_container.dart';
 
-abstract class IDependency<T> implements IBaseDependency<T> {
+abstract class IBlueprint<T> implements IBaseBlueprint<T> {
   T createValue(IContainer scope);
 
   bool shouldUpdateValue(T resolvedValue);